pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-team"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.44.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
search-api = [
    "tavily-python>=0.3.0",
]
output-extra = [
    "openpyxl>=3.1.0",
    "weasyprint>=62.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.5.0",
]

[project.scripts]
research-team = "research_team.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m 'not interactive and not e2e'"
markers = [
    "integration: integration tests requiring external services",
    "e2e: end-to-end tests requiring pi-agent and github-copilot login",
    "interactive: tests requiring a visible browser and manual interaction (skipped by default)",
]
